+++ 이지민/demo/src/main/java/com/todo/demo/repository/UserRepository.java
@@ -7,7 +7,7 @@ import com.todo.demo.model.User;
 import java.util.Optional;
 
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
-    Optional<User> findById(Long id);
+    Optional<User> findById(Integer id);
 }
